team manager integrated || update

+++ tp-teams-handling-api/tp-teams-handling-api/src/main/java/fr/sorbonne/paris/nord/university/tpteamshandlingapi/repository/TeamRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamRepository extends JpaRepository<TeamEntity, Integer> {
+public interface TeamRepository extends JpaRepository<TeamEntity, Long> {
 }
